refactor(app): separate reusable hooks from app.tsx
