+++ apps/renderer/src/hooks/useOpenFilePath.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+
+export function useOpenFilePath(loadFileInTab: (path: string) => Promise<void>) {
+  useEffect(() => {
+    window.api.onOpenFilePath((path) => {
+      void loadFileInTab(path);
+    });
+    return () => {
+      window.api.removeOpenFilePathListener();
+    };
+  }, [loadFileInTab]);
+}
